Правки по замечаним
