+++ js/util.js
@@ -47,30 +47,7 @@ function createNewRandomArray(array) {
   return shuffleArray().slice(getRandomIntegerInclusive(0,array.length - 1));
 }
 
-function showAlert(message, timeout) {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-  alertContainer.style.color = 'white';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, timeout);
-}
-
 export {
-  showAlert,
   getRandomIntegerInclusive,
   getRandomFloatInclusive,
   getRandomArrayElement,
